Small reorganization of files. Base of gui built. Logo added. Jfoenix and ikonli libraries added.
