+++ GRAPH-moment.iml
@@ -3,7 +3,9 @@
   <component name="NewModuleRootManager" inherit-compiler-output="true">
     <exclude-output />
     <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/test" isTestSource="true" />
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/assets" type="java-resource" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
@@ -17,5 +19,9 @@
         <SOURCES />
       </library>
     </orderEntry>
+    <orderEntry type="library" name="jfoenix-9.0.8" level="project" />
+    <orderEntry type="library" name="org.kordamp.ikonli:ikonli-materialdesign-pack:11.3.4" level="project" />
+    <orderEntry type="library" name="org.kordamp.ikonli:ikonli-fontawesome-pack:11.3.4" level="project" />
+    <orderEntry type="library" name="org.kordamp.ikonli:ikonli-javafx:11.3.4" level="project" />
   </component>
 </module>
