+++ src/com/github/mtdrewski/GRAPH_moment/main/view/Main.java
@@ -1,4 +1,4 @@
-package com.github.mtdrewski.GRAPH_moment.main;
+package com.github.mtdrewski.GRAPH_moment.main.view;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -10,10 +10,14 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("view/sample.fxml"));
+
+        Parent root = FXMLLoader.load(getClass().getResource("main.fxml"));
         primaryStage.setTitle("Hello World");
-        primaryStage.setScene(new Scene(root, 300, 275));
+        primaryStage.setScene(new Scene(root, 1200, 800));
+        primaryStage.setMinWidth(800);
+        primaryStage.setMinHeight(500);
         primaryStage.show();
+
     }
