+++ src/com/github/mtdrewski/GRAPH_moment/main/view/main.css
@@ -1,0 +1,32 @@
+#left-pane {
+    -fx-background-color: white;
+    -fx-padding: 100 0 0 0;
+}
+
+.left-pane-button {
+    
+    -fx-font-size: 18px;
+    -fx-alignment: baseline-left;
+
+    -fx-padding: 0 0 0 50px;
+    -fx-graphic-text-gap: 20px;
+
+    -fx-background-radius: 0;
+
+    -fx-min-width: 250px;
+    -fx-max-width: 250px;
+    -fx-min-height: 50px;
+    -fx-max-height: 50px;
+}
+
+.left-pane-button:hover {
+    -fx-background-color: #e1e1e1;
+}
+
+.graph-display-container {
+    -fx-background-color: #c2c2c2;
+}
+
+.graph-display {
+    -fx-background-color: dimgrey;
+}
