+++ src/com/github/mtdrewski/GRAPH_moment/main/view/main.fxml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXButton?>
+<?import java.net.URL?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.VBox?>
+<?import org.kordamp.ikonli.javafx.FontIcon?>
+
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.web.WebView?>
+<?import javafx.scene.shape.SVGPath?>
+<BorderPane id="scene" fx:id="root" minHeight="500.0" minWidth="800.0" prefHeight="800.0" prefWidth="1200.0"
+            xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="com.github.mtdrewski.GRAPH_moment.main.view.Main">
+    <stylesheets>
+        <URL value="@main.css"/>
+    </stylesheets>
+
+    <center>
+        <AnchorPane prefHeight="200.0" prefWidth="200.0" styleClass="graph-display-container"
+                    BorderPane.alignment="CENTER">
+            <children>
+                <!-- GRAPH DISPLAY GOES HERE -->
+                <!--<Canvas AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
+                        AnchorPane.topAnchor="0.0"/>-->
+            </children>
+        </AnchorPane>
+    </center>
+
+    <left>
+        <VBox id="left-pane" maxWidth="250" minWidth="250">
+            <JFXButton styleClass="left-pane-button" text="Export" VBox.vgrow="ALWAYS">
+                <graphic>
+                    <FontIcon iconLiteral="mdi-alarm" iconSize="30"/>
+                </graphic>
+            </JFXButton>
+            <JFXButton styleClass="left-pane-button" text="Export" VBox.vgrow="ALWAYS">
+                <graphic>
+                    <FontIcon iconLiteral="mdi-alarm" iconSize="30"/>
+                </graphic>
+            </JFXButton>
+            <JFXButton styleClass="left-pane-button" text="Export" VBox.vgrow="ALWAYS">
+                <graphic>
+                    <FontIcon iconLiteral="mdi-alarm" iconSize="30"/>
+                </graphic>
+            </JFXButton>
+            <JFXButton styleClass="left-pane-button" text="Export" VBox.vgrow="ALWAYS">
+                <graphic>
+                    <FontIcon iconLiteral="mdi-alarm" iconSize="30"/>
+                </graphic>
+            </JFXButton>
+            <JFXButton styleClass="left-pane-button" text="Export" VBox.vgrow="ALWAYS">
+                <graphic>
+                    <FontIcon iconLiteral="mdi-alarm" iconSize="30"/>
+                </graphic>
+            </JFXButton>
+        </VBox>
+    </left>
+</BorderPane>
