+++ src/com/github/mtdrewski/GRAPH_moment/main/view/sample.fxml
@@ -3,6 +3,41 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<GridPane fx:controller="com.github.mtdrewski.GRAPH_moment.main.controller.Controller"
-          xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10">
-</GridPane>
+
+
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import com.jfoenix.controls.JFXListView?>
+<?import javafx.scene.layout.FlowPane?>
+<?import com.jfoenix.controls.JFXButton?>
+<StackPane fx:id="root" fx:controller="com.github.mtdrewski.GRAPH_moment.main.controller.Controller"
+           xmlns:fx="http://javafx.com/fxml">
+
+<AnchorPane maxWidth="500">
+    <JFXListView fx:id="list1" styleClass="custom-jfx-list-view"
+                 layoutY="50">
+        <Label>Item 1</Label>
+        <JFXListView fx:id="subList" styleClass="sublist">
+            <groupnode>
+                <Label styleClass="sub-label">Item 2</Label>
+            </groupnode>
+            <Label>Item 2.1</Label>
+            <Label>Item 2.2</Label>
+        </JFXListView>
+        <Label>Item 3</Label>
+        <Label>Item 4</Label>
+        <Label>Item 5</Label>
+        <JFXListView fx:id="subList1" styleClass="sublist">
+            <groupnode>
+                <Label styleClass="sub-label">Item 6</Label>
+            </groupnode>
+            <Label>Item 6.1</Label>
+            <Label>Item 6.2</Label>
+        </JFXListView>
+        <Label>Item 7</Label>
+        <Label>Item 8</Label>
+        <Label>Item 9</Label>
+    </JFXListView>
+</AnchorPane>
+
+</StackPane>
